change int to size_t

+++ layer.hpp
@@ -9,11 +9,11 @@
 template <typename T>
 struct layer {
     std::function<T(T)> g;
-    int nodes;
+    size_t nodes;
     matrix<T> theta, a;
 
-    layer(int, int, std::function<T(T)>);
-    layer(int, int);
+    layer(size_t, size_t, std::function<T(T)>);
+    layer(size_t, size_t);
     layer();
     matrix<T> output(const matrix<T> &) const;
 };
@@ -23,14 +23,14 @@ template <typename T>
 layer<T>::layer() {}
 
 template <typename T>
-layer<T>::layer(int prv, int cur, std::function<T(T)> f) {
+layer<T>::layer(size_t prv, size_t cur, std::function<T(T)> f) {
     theta.resize(cur, prv + 1);
     nodes = cur;
     g = f;
 }
 
 template <typename T>
-layer<T>::layer(int prv, int cur) {
+layer<T>::layer(size_t prv, size_t cur) {
     theta.resize(cur, prv + 1);
     nodes = cur;
     g = [](const T &x) { return x; };
